python/invoke_cloud_workflows/google_cloud_run/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
ptvsd==4.3.2 # Required for debugging.
requests==2.25.1
google-cloud-secret-manager==2.6.0
google-cloud-storage==2.1.0
google-cloud-workflows==1.5.0
